+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stellar-agent-kit/core
 
+## 0.1.9
+
 ## 0.1.8
 
 ## 0.1.7